requirements.txt
Django==5.2.12
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
psycopg[binary]==3.2.12
redis==5.3.0
celery==5.5.3
django-celery-beat==2.8.1
stripe==12.5.1
python-dotenv==1.1.1
textblob==0.18.0.post0
scikit-learn==1.7.2
reportlab==4.4.2
pypdf==6.0.0
python-docx==1.2.0
